Sizes: INT instead of DINT/REAL, "Fill mm from inches" button
Fields max out well under 32k mm (~80 in), so INT (16-bit) is plenty —
swap DINT mm + REAL in for INT mm + INT in. ST_SizeSetpoints now packs
4×INT, half the size of before. Inches stop being decimal in the UI.

Conversion direction reversed: the dialog button now computes mm from
inches (mm = round(in × 25.4)), matching how operators usually quote
cardboard sizes in inches.

Existing rows in sizes.db still load — _row_to_size coerces width_in /
length_in to int, so REALs stored under the old schema round down.

PLC-side: update ST_SizeSetpoints UDT to 4×INT to match.

+++ sizes_store.py
@@ -36,8 +36,8 @@ CREATE TABLE IF NOT EXISTS cardboard (
     name       TEXT    NOT NULL,
     width_mm   INTEGER NOT NULL,
     length_mm  INTEGER NOT NULL,
-    width_in   REAL    NOT NULL,
-    length_in  REAL    NOT NULL,
+    width_in   INTEGER NOT NULL,
+    length_in  INTEGER NOT NULL,
     created_at TEXT    NOT NULL DEFAULT (datetime('now')),
     updated_at TEXT
 );
@@ -46,8 +46,8 @@ CREATE TABLE IF NOT EXISTS others (
     name       TEXT    NOT NULL,
     width_mm   INTEGER NOT NULL,
     length_mm  INTEGER NOT NULL,
-    width_in   REAL    NOT NULL,
-    length_in  REAL    NOT NULL,
+    width_in   INTEGER NOT NULL,
+    length_in  INTEGER NOT NULL,
     created_at TEXT    NOT NULL DEFAULT (datetime('now')),
     updated_at TEXT
 );
@@ -66,19 +66,20 @@ class Size:
     name:      str
     width_mm:  int
     length_mm: int
-    width_in:  float
-    length_in: float
+    width_in:  int
+    length_in: int
     id:        int | None = None       # None until persisted
 
 
 def _row_to_size(row: sqlite3.Row) -> Size:
+    # Old DBs may have stored width_in/length_in as REAL — coerce to int.
     return Size(
         id        = row["id"],
         name      = row["name"],
-        width_mm  = row["width_mm"],
-        length_mm = row["length_mm"],
-        width_in  = row["width_in"],
-        length_in = row["length_in"],
+        width_mm  = int(row["width_mm"]),
+        length_mm = int(row["length_mm"]),
+        width_in  = int(row["width_in"]),
+        length_in = int(row["length_in"]),
     )
